requirements.txt
pywebview
pix2tex[gui]
Pillow
pyperclip
torch
torchvision
flask